edits to plots
