+++ examples/grids/plot_log10Q.py
@@ -5,6 +5,7 @@ Create a model SED
 import matplotlib.pyplot as plt
 
 from synthesizer.grid import Grid
+from synthesizer.plots import plot_log10Q
 
 if __name__ == '__main__':
 
@@ -13,8 +14,10 @@ if __name__ == '__main__':
 
     grid = Grid(grid_name, grid_dir=grid_dir)
 
-    fig, ax = grid.plot_log10Q(ion='HI')
+    # plot grid of HI ionising luminosities
+    fig, ax = plot_log10Q(grid, ion='HI')
     plt.show()
 
-    fig, ax = grid.plot_log10Q(ion='HeII')
+    # plot grid of HeII ionising luminosities
+    fig, ax = plot_log10Q(grid, ion='HeII')
     plt.show()
